Split out getThemeDetailsUrl
