+++ client/state/themes/selectors/index.js
@@ -11,9 +11,7 @@ import config from 'config';
 import {
 	getCustomizerUrl,
 	getSiteSlug,
-	getSiteOption,
 	isJetpackSite,
-	canJetpackSiteManage,
 	hasJetpackSiteJetpackThemesExtendedFeatures,
 	isJetpackSiteMultiSite,
 } from 'state/sites/selectors';
@@ -44,38 +42,7 @@ export { isRequestingTheme } from 'state/themes/selectors/is-requesting-theme';
 export { isRequestingActiveTheme } from 'state/themes/selectors/is-requesting-active-theme';
 export { isWpcomTheme } from 'state/themes/selectors/is-wpcom-theme';
 export { isWporgTheme } from 'state/themes/selectors/is-wporg-theme';
-
-/**
- * Returns the URL for a given theme's details sheet.
- *
- * @param  {object}  state   Global state tree
- * @param  {string}  themeId Theme ID
- * @param  {?number} siteId  Site ID to optionally use as context
- * @returns {?string}         Theme details sheet URL
- */
-export function getThemeDetailsUrl( state, themeId, siteId ) {
-	if ( ! themeId ) {
-		return null;
-	}
-
-	if (
-		isJetpackSite( state, siteId ) &&
-		! (
-			config.isEnabled( 'manage/themes/details/jetpack' ) &&
-			canJetpackSiteManage( state, siteId ) &&
-			hasJetpackSiteJetpackThemesExtendedFeatures( state, siteId )
-		)
-	) {
-		return getSiteOption( state, siteId, 'admin_url' ) + 'themes.php?theme=' + themeId;
-	}
-
-	let baseUrl = oldShowcaseUrl + themeId;
-	if ( config.isEnabled( 'manage/themes/details' ) ) {
-		baseUrl = `/theme/${ themeId }`;
-	}
-
-	return baseUrl + ( siteId ? `/${ getSiteSlug( state, siteId ) }` : '' );
-}
+export { getThemeDetailsUrl } from 'state/themes/selectors/get-theme-details-url';
 
 /**
  * Returns the URL for a given theme's setup instructions
